requirements.txt
PyQt5>=5.15.0
pynput>=1.7.6
Pillow>=9.0.0
